Add txRow component to fix issues with Radium
